Add collision detection for map
Add collision detection for Map class. Checks if the given
bounding box will collide with a wall if it moves at the given
speed in the given direction. Protagonist will not move if the
movement causes collision with a wall. Guards will stop and turn
around if their next movement will collide with a wall.