Add start and game over screens
Prior to this commit, there were no separate start or game over
screens in the games. Now both screens are drawn and the logic
to implement them has been added.

NOTE: There is an option for Nux Mode on the start screen even
though it has not yet been implemented.